fix: correcciones de seguridad, tipos y formato de moneda para IA
- Fix corrupción monetaria en expenses_view.py: usar formateo seguro de Decimal
- Fix tipos en ai_model.py: variacion_total_pct y variacion_ticket_pct retornan Decimal
- Fix inyección SQL en expense_repository.py: construcción segura de queries
- Fix bypass de seguridad en base_repository.py: filtro por familia con familia_id=0
- Fix formato de moneda para IA en memory_event_handler.py: usar format_pesos_ai
- Documentación de correcciones en CORRECCIONES_2026-06-09.md

Todos los cambios verificados y tests pasando.

+++ models/ai_model.py
@@ -61,7 +61,7 @@ class CategoryMetric(BaseModel):
     ticket_anterior: Decimal | None = None
 
     @property
-    def variacion_total_pct(self) -> float | None:
+    def variacion_total_pct(self) -> Decimal | None:
         """Variación porcentual del gasto total vs mes anterior."""
         if self.total_anterior and self.total_anterior > 0:
             return (
@@ -70,7 +70,7 @@ class CategoryMetric(BaseModel):
         return None
 
     @property
-    def variacion_ticket_pct(self) -> float | None:
+    def variacion_ticket_pct(self) -> Decimal | None:
         """Variación porcentual del ticket promedio vs mes anterior."""
         if self.ticket_anterior and self.ticket_anterior > 0:
             return (
